requirements.txt
pypresence==4.3.0
psutil==5.9.8
pystray==0.19.5
Pillow==10.2.0
